EscrowSessionVault close/requestClose/withdraw implemented to EscrowSessionVaultDebugViewModel
