+++ wallet-sdk-core/src/androidMain/kotlin/com/michaeltchuang/walletsdk/core/railmpp/di/RailMppModule.android.kt
@@ -2,7 +2,8 @@ package com.michaeltchuang.walletsdk.core.railmpp.di
 
 import com.michaeltchuang.walletsdk.core.railmpp.data.repository.AndroidSessionVaultBalanceRepository
 import com.michaeltchuang.walletsdk.core.railmpp.domain.repository.SessionVaultBalanceRepository
-import com.michaeltchuang.walletsdk.core.railmpp.usecases.GetRemainingSessionVaultBalanceUseCase
+import com.michaeltchuang.walletsdk.core.railmpp.domain.usecases.MppWalletSignerUseCase
+import com.michaeltchuang.walletsdk.core.railmpp.domain.usecases.GetRemainingSessionVaultBalanceUseCase
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
@@ -11,4 +12,5 @@ val railMppModule =
     module {
         singleOf(::AndroidSessionVaultBalanceRepository) bind SessionVaultBalanceRepository::class
         singleOf(::GetRemainingSessionVaultBalanceUseCase)
+        singleOf(::MppWalletSignerUseCase)
     }
