--- wallet-sdk-ui/src/androidMain/kotlin/com/michaeltchuang/walletsdk/ui/settings/screens/EscrowSessionVaultDebugToolScreen.android.kt
@@ -1,10 +1,0 @@
-package com.michaeltchuang.walletsdk.ui.settings.screens
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
-import com.michaeltchuang.walletsdk.ui.test.EscrowSessionVaultDebugToolScreen
-
-@Composable
-actual fun EscrowSessionVaultDebugToolScreen(navController: Any) {
-    EscrowSessionVaultDebugToolScreen(navController = navController as NavController)
-}
