--- wallet-sdk-ui/src/commonMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/utils/PaymentUtil.kt
@@ -1,7 +1,0 @@
-package com.michaeltchuang.walletsdk.ui.liquidStream.utils
-
-expect fun startSettlePayment(
-    viewerAddress: String,
-    creatorAddress: String,
-    viewerAuthSignKey: ByteArray,
-)
