+++ wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/base/di/UiPlatformModule.ios.kt
@@ -1,7 +1,7 @@
 package com.michaeltchuang.walletsdk.ui.base.di
 
 import com.michaeltchuang.walletsdk.ui.liquidStream.IosViewerPaymentOrchestrator
-import com.michaeltchuang.walletsdk.ui.test.IosPaymentTestViewModel
+import com.michaeltchuang.walletsdk.ui.settings.viewmodels.EscrowSessionVaultDebugViewModel
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.module.dsl.viewModel
@@ -10,11 +10,8 @@ import org.koin.dsl.module
 actual fun uiPlatformModule(): Module =
     module {
         viewModel {
-            IosPaymentTestViewModel(
-                getLocalAccount = get(),
-                getAlgo25SecretKey = get(),
-                getFalcon24SecretKey = get(),
-                getHdSeed = get(),
+            EscrowSessionVaultDebugViewModel(
+                mppWalletSignerUseCase = get(),
             )
         }
         singleOf(::IosViewerPaymentOrchestrator)
